Add a custom alert for app insights
This allows the ability to set a custom KQL query
per project and trigger an alert based on that
query - for now we just have one use case, which
is to alert on an event in the
`dtos-manage-breast-screening/notifications` repo.
This could be adapted for future use by e.g.
creating multiple alerts for multiple queries.