Add an Alert on Exceptions for App Insights (#256)
We want to trigger an alert whenever there is an
exception for manage-breast-screening -
specifically for the Invite team's container app
jobs in the notifications app.

This should trigger whenever the Exception count
is above 0 - may want to tweak that but in theory
there should never be an Exception unless
something has completely failed.