DTOSS-11646: Setup Container Apps container probes
We need to configure a container probe (liveness/readiness & startup) for our Azure Container Apps deployment.
By default, the web application should only have a heath probe if the variable probe_path is set.
If probe_path is null then it's disabled.
Please note that we cannot do this for the Postgres DB container as there is no Django / restful interface to implement.
Also update the terradocs on the container app module and update the README.md for instructions on how to deploy probes from the root module.